refac: relocate metadata module
