+++ benchmark_analyzer/src/benchmark/group/element/mod.rs
@@ -3,12 +3,13 @@
 //!
 
 pub mod input;
+pub mod metadata;
 pub mod selector;
 
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::benchmark::metadata::Metadata;
+use metadata::Metadata;
 
 ///
 /// The benchmark element.
