+++ benchmark_analyzer/src/benchmark/mod.rs
@@ -4,7 +4,6 @@
 
 pub mod format;
 pub mod group;
-pub mod metadata;
 
 use std::collections::BTreeMap;
 use std::path::PathBuf;
